requirements.txt
flask
openai
Flask-Cors
bcrypt
google-cloud-firestore
smtplib
python-dotenv